Simulator ready to work
